+++ include/rendering/material.h
@@ -9,21 +9,21 @@ namespace Rendering
 class Material
 {
 public:
-	Material(const Math::Vector3D &color = {}, const double reflection = 0, const double refraction = 0);
+	Material(const Math::Vector3D &color = {}, const float reflection = 0, const float refraction = 0);
 	
 	void setColor(const Math::Vector3D &color);
 	const Math::Vector3D &color() const;
 	
-	void setReflection(const double reflection);
-	double reflection() const;
+	void setReflection(const float reflection);
+	float reflection() const;
 	
-	void setRefraction(const double refraction);
-	double refraction() const;
+	void setRefraction(const float refraction);
+	float refraction() const;
 	
 private:
 	Math::Vector3D _color;
-	double _reflection;
-	double _refraction;
+	float _reflection;
+	float _refraction;
 };
 
 } // namespace Rendering
