+++ src/math/vector3d.cpp
@@ -1,54 +1,55 @@
 #include "vector3d.h"
 
 #include <cmath>
+#include <emmintrin.h>
 #include <sstream>
 
 namespace Math
 {
 
-Vector3D::Vector3D(const double x, const double y, const double z, const double w) :
+Vector3D::Vector3D(const float x, const float y, const float z, const float w) :
 	_coordinates({x, y, z, w})
 {
 }
 
-double Vector3D::x() const
+float Vector3D::x() const
 {
 	return this->_coordinates[0];
 }
 
-void Vector3D::setX(const double x)
+void Vector3D::setX(const float x)
 {
 	this->_coordinates[0] = x;
 }
 
-double Vector3D::y() const
+float Vector3D::y() const
 {
 	return this->_coordinates[1];
 }
 
-void Vector3D::setY(const double y)
+void Vector3D::setY(const float y)
 {
 	this->_coordinates[1] = y;
 }
 
-double Vector3D::z() const
+float Vector3D::z() const
 {
 	return this->_coordinates[2];
 }
 
-void Vector3D::setZ(const double z)
+void Vector3D::setZ(const float z)
 {
 	this->_coordinates[2] = z;
 }
 
-double Vector3D::magnitude() const
+float Vector3D::magnitude() const
 {
-	return std::pow((std::pow(this->_coordinates[0], 2) + std::pow(this->_coordinates[1], 2) + std::pow(this->_coordinates[2], 2)), 0.5);
+	return std::pow((std::pow(this->_coordinates[0], 2.0f) + std::pow(this->_coordinates[1], 2.0f) + std::pow(this->_coordinates[2], 2.0f)), 0.5f);
 }
 
 Vector3D &Vector3D::normalize()
 {
-	double magnitude = this->magnitude();
+	float magnitude = this->magnitude();
 	
 	this->_coordinates[0] /= magnitude;
 	this->_coordinates[1] /= magnitude;
@@ -88,9 +89,9 @@ Vector3D Vector3D::coordinateProduct(const Vector3D &other) const
 	return returnValue;
 }
 
-double Vector3D::cos(const Vector3D &left, const Vector3D &right)
+float Vector3D::cos(const Vector3D &left, const Vector3D &right)
 {
-	double returnValue = 0;
+	float returnValue = 0;
 	
 	returnValue = ((left * right) / (left.magnitude() * right.magnitude()));
 	
@@ -115,7 +116,7 @@ Vector3D &Vector3D::operator-=(const Vector3D &other)
 	return *this;
 }
 
-Vector3D &Vector3D::operator*=(const double scalar)
+Vector3D &Vector3D::operator*=(const float scalar)
 {
 	this->_coordinates[0] *= scalar;
 	this->_coordinates[1] *= scalar;
@@ -124,7 +125,7 @@ Vector3D &Vector3D::operator*=(const double scalar)
 	return *this;
 }
 
-Vector3D &Vector3D::operator/=(const double scalar)
+Vector3D &Vector3D::operator/=(const float scalar)
 {
 	this->_coordinates[0] /= scalar;
 	this->_coordinates[1] /= scalar;
@@ -138,12 +139,12 @@ Vector3D Vector3D::operator-() const
 	return *this * (-1.0);
 }
 
-double &Vector3D::operator[](const size_t index)
+float &Vector3D::operator[](const size_t index)
 {
 	return this->_coordinates[index];
 }
 
-double Vector3D::operator[](const size_t index) const
+float Vector3D::operator[](const size_t index) const
 {
 	return this->_coordinates[index];
 }
@@ -176,9 +177,9 @@ Vector3D operator-(const Vector3D &left, const Vector3D &right)
 	return returnValue;
 }
 
-double operator*(const Vector3D &left, const Vector3D &right)
+float operator*(const Vector3D &left, const Vector3D &right)
 {
-	double returnValue = 0;
+	float returnValue = 0;
 	
 	returnValue += (left._coordinates[0] * right._coordinates[0]);
 	returnValue += (left._coordinates[1] * right._coordinates[1]);
@@ -187,7 +188,7 @@ double operator*(const Vector3D &left, const Vector3D &right)
 	return returnValue;
 }
 
-Vector3D operator*(const Vector3D &left, const double right)
+Vector3D operator*(const Vector3D &left, const float right)
 {
 	Vector3D returnValue = left;
 	
@@ -196,12 +197,12 @@ Vector3D operator*(const Vector3D &left, const double right)
 	return returnValue;
 }
 
-Vector3D operator*(const double left, const Vector3D right)
+Vector3D operator*(const float left, const Vector3D right)
 {
 	return right * left;
 }
 
-Vector3D operator/(const Vector3D &left, const double right)
+Vector3D operator/(const Vector3D &left, const float right)
 {
 	Vector3D returnValue = left;
 	
@@ -210,7 +211,7 @@ Vector3D operator/(const Vector3D &left, const double right)
 	return returnValue;
 }
 
-Vector3D operator/(const double left, const Vector3D right)
+Vector3D operator/(const float left, const Vector3D right)
 {
 	return right / left;
 }
